added mobility to arcgis interface
