--- cea/GUI/mobility_tool.py
@@ -1,47 +1,0 @@
-import cea
-from cea.interfaces.arcgis.CityEnergyAnalyst import add_message
-
-__author__ = "Daren Thomas"
-__copyright__ = "Copyright 2016, Architecture and Building Systems - ETH Zurich"
-__credits__ = ["Daren Thomas"]
-__license__ = "MIT"
-__version__ = "0.1"
-__maintainer__ = "Daren Thomas"
-__email__ = "cea@arch.ethz.ch"
-__status__ = "Production"
-
-
-class MobilityTool(object):
-    def __init__(self):
-        self.label = 'Emissions Mobility'
-        self.description = 'Calculate emissions and primary energy due to mobility'
-        self.canRunInBackground = False
-
-    def getParameterInfo(self):
-        import arcpy
-        scenario_path = arcpy.Parameter(
-            displayName="Path to the scenario",
-            name="scenario_path",
-            datatype="DEFolder",
-            parameterType="Required",
-            direction="Input")
-
-        return [scenario_path]
-
-    def isLicensed(self):
-        return True
-
-    def updateParameters(self, parameters):
-        return
-
-    def updateMessages(self, parameters):
-        return
-
-    def execute(self, parameters, messages):
-        from cea.analysis.mobility import lca_mobility
-
-        scenario_path = parameters[0].valueAsText
-        gv = cea.globalvar.GlobalVariables()
-        gv.log = add_message
-        locator = cea.inputlocator.InputLocator(scenario_path=scenario_path)
-        lca_mobility(locator=locator)
